Covariates Generation

ContextualBandits.CovariatesGeneration.CovariatesIndependent — Type

CovariatesIndependent <: Sampleable{Multivariate,Continuous}
CovariatesIndependent(marginals, intercept=true, reduce_category=true)

Sampleable to randomly generate a vector of covariates using marginal distributions assuming the values are independent of each other.

Arguments

marginals::Array{Distribution{Univariate,S} where S<:ValueSupport,1}: vector of univariate distributions from the Distributions package

intercept=true: include an intercept term in the vector of covariates

reduce_category=true: omit the first dummy variable for Categorical covariates

Examples

using Distributions
FX = CovariatesIndependent([Categorical([1/3,1/3,1/3]),Normal(0,1)])
rand(FX)

ContextualBandits.CovariatesGeneration.CovariatesCopula — Type

CovariatesCopula{C} <: Sampleable{Multivariate,Continuous}
CovariatesCopula(marginals,copula,intercept=true,reduce_category=true)

Sampleable to randomly generate a vector of covariates using marginal distributions and copulas from the Copulas.jl package.

Categorical marginal distributions are assumed unordered (nominal) and automatically generated as a set of dummy variables. To use ordered categorical variables, use the OrdinalDiscrete type.

Arguments

marginals::Vector{Distribution{Univariate,S} where S<:ValueSupport}: vector of univariate distributions from the Distributions package

copula: copula type from the Copulas package

intercept=true: include an intercept term in the vector of covariates

reduce_category=true: omit the first dummy variable for Categorical covariates

Examples

using Distributions, Copulas
copula = GaussianCopula([1 0.2; 0.2 1])
FX = CovariatesCopula([Categorical([1/3,1/3,1/3]),Normal(0,1)],copula)
rand(FX)

ContextualBandits.CovariatesGeneration.OrdinalDiscrete — Type

struct OrdinalDiscrete{T<:Real,P<:Real,Ts<:AbstractVector{T},Ps<:AbstractVector{P}} <: DiscreteUnivariateDistribution

Type to sample from a discrete distribution. Unlike DiscreteNonParametric or Categorical, this type is interpreted as ordinal by CovariatesCopula and CovariatesIndependent.

WARNING: although this type is a Distribution, I have only defined the rand and quantile functions, which are required for covariate generation.

The implementation wraps a DiscreteNonParametric variable in a new structure.