Model

There are many "flavors" of bandit models and this package is not comprehensive. Therefore, users must understand the underlying model for which this package was designed. The contextual bandit model that we present here was formally introduced in the paper Alban A, Chick SE, Zoumpoulis SI (2024) Learning Personalized Treatment Strategies with Predictive and Prognostic Covariates in Adaptive Clinical Trials. We introduce some terminology that will be used throughout this documentation.

A trial is a sequence of allocations of T subjects to one of n alternative treatments (arms). Before allocating each subject, we observe a context, which is a vector with m covariates. After each allocation of a subject, we observe the outcome, a noisy scalar that represents how effective the treatment is for the subject (higher values are better). By the end of the trial, we have gathered the following data:

W: a T-dimensional vector of integers in 1:n of the treatments allocated to each subject.

X: a mxT-dimensional matrix where each column is the vector of covariates for a subject. Notice that Julia stores arrays in column-major order.

Y: a T-dimensional vector of scalars with the outcome of each subject.

Policy

A policy makes allocation decisions: it decides to which treatment each subject is allocated. A policy is represented by a subtype of Policy:

ContextualBandits.Policy — Type

Policy

Supertype for contextual bandit policies.

The functions initialize!, state_update!, allocation, and implementation take subtypes of Policy as the first argument. Each subtype of Policy should implement these functions.

The main method for any subtype of policy is allocation:

ContextualBandits.allocation — Method

allocation(policy::Policy,Xcurrent,W,X,Y[, rng])

Return a treatment to allocate a patient with covariates Xcurrent, given that the trial has observed W, X, and Y. W is the vector of treatments, X is the matrix of covariates, and Y is the vector of outcomes.

The dimension of Y may be smaller than that of W and X because of delays in outcomes.

The allocation depends only on the available data W,X,Y when the allocation is made. However, the policy can also maintain a state. A state can be useful for computational efficiency. For instance, a linear regression model can be updated sequentially rather than retrained for every allocation. To maintain a state, the policy is first initialized and then updated after every observation:

ContextualBandits.initialize! — Method

initialize!(policy::Policy[, W, X, Y])

Initialize the state of a policy before a trial starts. W, X, and Y is data collected in a pilot that can be used to initialize the policy. W is the vector of treatments, X is the matrix of covariates, and Y is the vector of outcomes.

ContextualBandits.state_update! — Method

state_update!(policy::Policy,W,X,Y[, rng])

Update the state of a policy given the data W, X, and Y. W is the vector of treatments, X is the matrix of covariates, and Y is the vector of outcomes.

For example, the policy may do Bayesian updating to get posterior parameters.

A policy may also define an implementation that determines which treatment a subject arriving post-trial (once the trial has concluded and no more data is being gathered):

ContextualBandits.implementation — Method

implementation(policy::Policy,X_post,W,X,Y)

Implement a treatment for covariates X_post given that the trial observed W, X, and Y. W is the vector of treatments, X is the matrix of covariates, and Y is the vector of outcomes.

In Policies, we present the policies provided by this package. You can define your own policy by creating a subtype of Policy that defines methods for the above functions.

Covariates

The covariates X are random draws from a distribution FX, which are independent between subjects. The package provides CovariatesIndependent and CovariatesCopula, which are described in Covariates Generation, to randomly generate covariates. As an example, a covariate that follows a standard normal distribution can be generated as follows:

using ContextualBandits
using Distributions

FX = CovariatesIndependent([Normal()])
X = rand(FX)

Outcomes

The outcomes of a trial, Y, are generated by an outcome model:

ContextualBandits.OutcomeModel — Type

OutcomeModel

Supertype for models that determine mean outcomes and noisy outcomes.

The functions outcome_model_state!, mean_outcome, noisy_outcome, and noise_outcome take subtypes of OutcomeModel as the first argument. Each subtype should implement these functions.

ContextualBandits.mean_outcome — Function

mean_outcome(outcome_model::OutcomeModel,W,X)

Compute the mean outcome of treatment W with covariates X.

ContextualBandits.noise_outcome — Function

noise_outcome(outcome_model::OutcomeModel,rng::AbstractRNG=Random.GLOBAL_RNG)

Generate noise object to be passed to noisy_outcome.

It can generate a normal distribution with zero mean for Gaussian models or vectors of random variables for more complex models.

ContextualBandits.noisy_outcome — Function

noisy_outcome(outcome_model::OutcomeModel,W,X,Z)

Compute noisy outcome of treatment W with covariates X given noise Z. Z is generated with noise_outcome.

Usually calls mean_outcome and adds zero mean noise Z.

ContextualBandits.outcome_model_state! — Function

outcome_model_state!(outcome_model::OutcomeModel,rng::AbstractRNG=Random.GLOBAL_RNG)

Set the state of an outcome model. Mainly used to change the state of random instances.

In Outcome Models, we present the outcome models provided by this package. You can define your own model by creating a subtype of OutcomeModel that defines methods for the above functions.

Delay

The model allows for fixed delays in observing the outcomes. simulation_stochastic accepts delay as an integer parameter that determines the number of subjects that are allocated before the outcome of a subject is observed. For example if delay=3, then the outcome of subject 1 will be observed only after the allocation of subject 4.

Simulation of a Trial

The model can be understood with the following code snippet that broadly illustrates how the different components enter in a simulation. A simulation will perform this code for many replications, record metrics, and summarize the metrics. See Simulation for more details on the simulation functions.

# ...
# loop over subjects in trial
initialize!(policy)
for t in 0:(T+delay)
    # Allocate
    if 1 <= t <= T
        # Covariates of the newly arrived subject that has not yet been treated
        Xcurrent = rand(FX)
        w = allocation(policy,Xcurrent,view(W,1:(t-1)),view(X,:,1:(t-1)),view(Y,1:(t-delay-1)),rng)
        @assert w in 1:n "The policy $(typeof(policy)) made an allocation outside the range 1-$n"
        X[:,t] = Xcurrent
        W[t] = w
        Y[t] = noisy_outcome(outcome_model,w,view(X,:,t),Z[t])
    end

    # Available data at time t
    Wav = view(W,1:min(t,T))
    Xav = view(X,:,1:min(t,T))
    Yav = view(Y,1:(t-delay))

    # Update state of policy
    state_update!(policy,Wav,Xav,Yav,rng)
end

Metrics and performance

We distinguish between two types of metrics: online and offline. Online performance metrics refer to performance associated with the subjects allocated in the trial, while offline refers to the outcome of subjects that would be allocated to treatment post-trial once an implementation has been chosen. Regret for a subject with covariates x allocated to treatment w is defined as follows:

regret = maximum([mean_outcome(outcome_model, wmax, x) - mean_outcome(outcome_model, w, x) for wmax in 1:n])

For subjects allocated in the trial, we refer to it as online regret, and for subjects that received the treatment as the post-trial implementation, we refer to it as offline regret. In Alban, Chick, Zoumpoulis (2024) offline regret is referred to as Expected Opportunity Cost (EOC).

Another important metric is Probability of Incorrect Selection (PICS):

PICS = regret > 0

Note

This package assumes that larger values of the output are better when computing regret and for designing policies.

More detail is provided in the output of the Simulation.

Linear Model with Labeling

A common way to model the the outcomes as a function of the covariates and treatment is through linear model. Most effective policies require a statistical model that learns the coefficients of the linear model. The vast majority of implemented policies in this package use Bayesian Linear Regression to learn those coefficients. We use mu as the true coefficients of the model that are generating the outcomes. theta is the prior/posterior estimate of mu and Sigma is the prior/posterior covariance matrix of the Bayesian linear regression.

A reasonable way to implement a linear model is to model each treatment separately. However, we use a more flexible framework of interaction between treatment and covariates with a labeling (see Labeling for details). Thus, the final model is Y = interact(W,n,X,labeling) * mu + sigma * randn(), where sigma^2 is the standard deviation of the noise. Alban, Chick, Zoumpoulis (2024) uses the operator $\otimes$ to represent the interact function.

ContextualBandits.interact — Function

interact(w, n, x[, labeling])

Interact a treatment w among n treatment alternatives with covariate vector x (m=length(x)).

When labeling is not provided, the function returns a vector WX of length m*n, where all entries are zeros except for WX[(w-1)*m+1:w*m] = x. WX is split into n blocks of size m, where the w-th block is equal to x.

When labeling is provided, the function returns a vector WX of length sum(labeling). labeling is a boolean vector of length (n+1)*m that indicates which covariates are predictive and which are prognostic. The first m entries correspond to prognostic terms, the following m entries correspond to the terms predictive with respect to treatment 1, then m entries for treatment 2, and so on.

If w is a vector, then x must be a matrix such that length(w)==size(x,2), and the output is a matrix with sum(labeling) rows and length(w) columns.

See also: interact!

#Examples

julia> w = 2;
       n = 2;
       x = [1,3,4]
       interact(w, n, x)
Vector{Float64}:
 0.0
 0.0
 0.0
 1.0
 3.0
 4.0

julia> w = [1,2]
       n = 2
       x = [1 1;
            2 3;
            5 4]
       labeling = Bool.([1,1,0,0,0,1,0,1,0])
       interact(w,n,x,labeling)
4×2 Matrix{Float64}:
 1.0  1.0
 2.0  3.0
 5.0  0.0
 0.0  3.0

ContextualBandits.interact! — Function

interact!(WX, w, n, x[, labeling])

In-place version of interact.

Labeling

A labeling "labels" how each covariate interacts with each treatment. It is a boolean vector of size (n+1)*m split into n+1 blocks of m entries. The first m entries represent prognostic covariates, those covariates that have an equal effect on all treatments. By default, all prognostic labels are false, meaning there are no prognostic effects. The remaining n blocks represent the predictive covariates with respect to the corresponding treatment, those covariates that have a treatment-specific effect. By default, all predictive labels are true, meaning each covariates has an effect specific to each treatment, or equivalently, each treatment has a separate model.

Generally, the first covariate is an intercept term that is 1 for all subjects. Thus, the first prognostic term represents an intercept term, and the first term for each block of predictive covariates are treatment effects.

Bayesian linear regression

We use the following type to learn Bayesian linear regression models. Policies will often use this type

ContextualBandits.BayesLinearRegression — Type

BayesLinearRegression(n, m, theta0, Sigma0, sample_std[, labeling])

Bayesian linear regression model with n treatments and m covariates.

See also: ContextualBandits.initialize!, ContextualBandits.state_update!, BayesUpdateNormal, BayesUpdateNormal!

This type needs to be first initialized and then data can be passed sequentially (or in batches) to update the posterior distribution.

ContextualBandits.initialize! — Method

initialize!(model::BayesLinearRegression)

Reset the posterior mean model.theta_t and covariance matrix model.Sigma_t to the prior mean theta0 and covariance matrix Sigma0.

ContextualBandits.state_update! — Method

state_update!(model::BayesLinearRegression, W, X, Y)

Update the posterior mean model.theta_t and covariance matrix model.Sigma_t after observing treatments W, covariates X, and outcomes Y.

To select a prior, the package provides the following utilities to generate the robust prior of Alban, Chick, Zoumpoulis (2024).

ContextualBandits.default_prior_linear — Function

default_prior_linear(n, m, sigma0, psi[, D, labeling])

The default prior is a method to specify a prior for a linear model that only depends on a few parameters. It returns the prior mean and covariance matrix for a linear model with n treatments and m covariates.

The prior mean is a vector of zeros.

The prior covariance matrix (Sigma0) has sigma0^2 in all entries of the diagonal. psi is the decay parameter for the covariance between coefficients of the same covariate and D is a symmetric distance matrix between treatments, which is of size (n,n). Off-diagonal elements of the covariance matrix that represent the covariance between two predictive coefficient for the same covariate with respect to two different treatments, say w1 and w2, are given by sigma0^2*exp(-psi*D[w1,w2]).

Example

n = 3
m = 3
sigma0 = 1.0
psi = log(2)
D = [0 1 2;
    1 0 1;
    2 1 0]
labeling = BitVector([0, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0])
theta0, Sigma0 = default_prior_linear(n, m, sigma0, psi, D, labeling)

ContextualBandits.robustify_prior_linear! — Function

robustify_prior_linear!(theta0, Sigma0, n, m[, labeling, z_alpha, c])

Robustify the prior mean vector theta0 and covariance matrix Sigma0 for a linear model with n treatments and m covariates. theta0 elements that are predictive are artificially increased by a factor of z_alpha times the square root of the largest element of Sigma0. Sigma0 is multiplied by scalar c.

By default z_alpha=2 and c=4.

Additional utilities

ContextualBandits.BayesUpdateNormal — Function

BayesUpdateNormal(theta,Sigma,X,y,sample_std)

Update Bayesian hyperparameters theta (mean vector) and Sigma (covariance matrix) of a linear regression model with regressor matrix X, outputs y, and sampling standard deviation sample_std.

Return a copy of the updated theta and Sigma.

X can be a vector of regressors or a matrix of regressors, where each column is a vector of regressors. In the latter case, y and sample_std can be vectors of the same length as the number of columns of X.

See also: BayesUpdateNormal!

ContextualBandits.BayesUpdateNormal! — Function

BayesUpdateNormal!(theta,Sigma,X,y,sample_std)

In-place version of BayesUpdateNormal