Outcome Models

In Model, we introduced the OutcomeModel abstract type. Here, we provide subtypes for linear outcome model.

ContextualBandits.OutcomeLinear — Type

OutcomeLinear <: OutcomeModel
OutcomeLinear(n, m, mu, sample_std[, labeling])

Linear outcome model with fixed coefficients mu. n is the number of treatments, m is the number of covariates, including the intercept term. The first m coefficients represent prognostic effects (effects independent of the treatment) for each of the m covariates. The following m coefficients represent the predictive effects of treatment 1 (interaction between treatment and covariates). In total, there are at most (n+1)*m coefficients. labeling is a boolean vector of length (n+1)*m indicationg active coefficients. If any coefficient is knwon to be zero (inactive), the corresponding entry of labeling should be false. The length of mu should be equal to the number of active coefficients=sum(labeling). By default, only predictive coefficients are active.

Outcomes are observed with white noise around the mean with sample standard deviation sample_std.

ContextualBandits.OutcomeLinearBayes — Type

OutcomeLinearBayes <: OutcomeModel
OutcomeLinearBayes(n, m, theta0, Sigma0, sample_std[, labeling])

Linear outcome model with coefficients randomly drawn from a normal prior distribution with mean vector theta0 and covariance matrix Sigma0. n is the number of treatments, m is the number of covariates, including the intercept term. The first m coefficients represent prognostic effects (effects independent of the treatment) for each of the m covariates. The following m coefficients represent the predictive effects of treatment 1 (interaction between treatment and covariates). In total, there are at most (n+1)*m coefficients. labeling is a boolean vector of length (n+1)*m indicationg active coefficients. If any coefficient is knwon to be zero (inactive), the corresponding entry of labeling should be false. The length of theta0 and the dimensions of Sigma0 should be equal to the number of active coefficients=sum(labeling). By default, only predictive coefficients are active.

Outcomes are observed with white noise around the mean with sample standard deviation sample_std.