Policies

In Model, we introduced the Policy abstract type. Here, we first introduced PolicyLinear that serves as the supertype for the policies with a linear model. Then, we introduce the concrete policies that are defined by the package. More detail about the policies defined here is provided in Alban, Chick, Zoumpoulis (2024).

Linear Policies

ContextualBandits.PolicyLinear — Type

PolicyLinear <: Policy

Abstract supertype that updates using the model with a labeling and implements the treatments strategy that maximizes expected outcomes. It does not provide an allocation policy, which should be defined for each subtype.

All subtypes must include a model::BayesLinearRegression field. initialize! and state_update! methods are defined to maintain the state. An implementation method is defined to implement the treatment with the largest expected value. An allocation method must be defined by the subtypes.

Expected Value of Information (EVI)

ContextualBandits.fEVI_MC_PolicyLinear — Type

fEVI_MC_PolicyLinear <: PolicyLinear
fEVI_MC_PolicyLinear(n, m, theta0, Sigma0, sample_std, FXtilde, etaon, etaoff, labeling=vcat(falses(m),trues(n*m)))

Allocate treatment using the fEVI-MC allocation policy and update based on the linear model with labeling to make an implementation.

ContextualBandits.fEVIDiscrete — Type

fEVIDiscrete <: PolicyLinearDiscrete
fEVIDiscrete(n, m, theta0, Sigma0, sample_std, FX, labeling=vcat(falses(m),trues(n*m)))

Allocate treatment using the fEVI allocation policy and update based on the linear model with labeling to make an implementation.

ContextualBandits.fEVIDiscreteOnOff — Type

fEVIDiscreteOnOff <: PolicyLinearDiscrete
fEVIDiscreteOnOff(n, m, theta0, Sigma0, sample_std, FX, P, T, labeling=vcat(falses(m),trues(n*m)))

Allocate treatment using the fEVI allocation policy (with online and offline rewards) and update based on linear model with labeling to make an implementation.

Thompson sampling

ContextualBandits.TSPolicyLinear — Type

TSPolicyLinear <: PolicyLinear
TSPolicyLinear(n, m, theta0, Sigma0, sample_std[, labeling])

Allocate treatment using Thompson Sampling and update based on the linear model with labeling to make an implementation.

ContextualBandits.TTTSPolicyLinear — Type

TTTSPolicyLinear <: PolicyLinear
TTTSPolicyLinear(n, m, theta0, Sigma0, sample_std, beta, maxiter[, labeling])

Allocate treatment using Top-Two Thompson Sampling and update based on the linear model with labeling to make an implementation.

Russo D (2020) Simple Bayesian algorithms for best arm identification. Operations Research 68(6)

Optimal Computing Budget Allocation (OCBA)

ContextualBandits.OCBAPolicyLinear — Type

OCBAPolicyLinear <: PolicyLinear
OCBAPolicyLinear(n, m, theta0, Sigma0, sample_std, predictive, labeling=vcat(falses(m),trues(n*m)))
OCBAPolicyLinear(n, m, theta0, Sigma0, sample_std, FX::Union{CovariatesCopula, CovariatesIndependent}, labeling=vcat(falses(m),trues(n*m)))

Allocate treatment using the OCBA algorithm and update based on the linear model with labeling to make an implementation.

predictive is a vector of integers that specifies the covariates that define predictive groups, i.e., patients in the same predictive group have the same covariate values for the covariates specified in predictive. Instead of predictive, you can pass FX, which is a CovariatesCopula or CovariatesIndependent object, and the predictive groups will be automatically determined based on the labeling.

NOTE: the covariates specified in predictive should be discrete. If it is not discrete, the algorithm will run but the results may not be meaningful.

Biased coin

ContextualBandits.BiasedCoinPolicyLinear — Type

BiasedCoinPolicyLinear <: PolicyLinear
BiasedCoinPolicyLinear(n, m, theta0, Sigma0, sample_std, predictive, prognostic, labeling=vcat(falses(m),trues(n*m)); p=vcat(0.5, 0.5*ones(n-1)/(n-1)), weights=vcat(0.0,ones(length(prognostic))/length(prognostic),0.0), target_fraction=ones(n)/n)
BiasedCoinPolicyLinear(n, m, theta0, Sigma0, sample_std, FX::Union{CovariatesCopula,CovariatesIndependent}, labeling=vcat(falses(m),trues(n*m)); p=vcat(0.5, 0.5*ones(n-1)/(n-1)), weights=nothing, target_fraction=ones(n)/n)

A policy that allocates treatments according to a biased coin design and implements using a linear model with a labeling.

See Zhao, W., Ma, W., Wang, F., & Hu, F. (2022). Incorporating covariates information in adaptive clinical trials for precision medicine. Pharmaceutical Statistics, 21(1), 176-195. for the measure of the score. However, this policy allocates using a fixed target_fraction, instead of a response-adaptive allocation. See RABC_OCBA_PolicyLinear for a response-adaptive version of this policy.

ContextualBandits.RABC_OCBA_PolicyLinear — Type

RABC_OCBA_PolicyLinear <: PolicyLinear
RABC_OCBA_PolicyLinear(n, m, theta0, Sigma0, sample_std, predictive, prognostic, labeling=vcat(falses(m),trues(n*m)); p=vcat(0.5, 0.5*ones(n-1)/(n-1)), weights=vcat(0.0,ones(length(prognostic))/length(prognostic),0.0))
RABC_OCBA_PolicyLinear(n, m, theta0, Sigma0, sample_std, FX::Union{CovariatesCopula,CovariatesIndependent}, labeling=vcat(falses(m),trues(n*m)); p=vcat(0.5, 0.5*ones(n-1)/(n-1)), weights=nothing)

Reasponse-Adaptive Biased Coin (RABC) using OCBA for the target fraction. A policy that allocates treatments according to a response-adative biased coin design and implements using a linear model with a labeling. The response-adaptive version target fraction is updated using the OCBA method.

See Zhao, W., Ma, W., Wang, F., & Hu, F. (2022). Incorporating covariates information in adaptive clinical trials for precision medicine. Pharmaceutical Statistics, 21(1), 176-195. for the measure of the score.

See BiasedCoinPolicyLinear for a non-response-adaptive version of this policy.

See OCBAPolicyLinear for the OCBA policy without biased coin.

Policy Modifiers

Infer labeling

ContextualBandits.InferLabelingPolicy — Type

InferLabelingPolicy{T<:PolicyLinear, S<:LabelingSelector} <: Policy
InferLabelingPolicy(subpolicy,selector,schedule,labeling0=trues((subpolicy.model.n+1)*subpolicy.model.m),sigma0=1e6,psi=1,D=fill(Inf,subpolicy.model.n,subpolicy.model.n),z_alpha=2,c=4)

Create a policy that modifies subpolicy by using a labeling selector to infer the labeling at a specified schedule.

Discretize policy

ContextualBandits.DiscretizePolicy — Type

DiscretizePolicy{T<:Policy} <: Policy
DiscretizePolicy(subpolicy::T, FX::Union{CovariatesIndependent,CovariatesCopula}, breakpoints) where {T<:Policy}

Modify subpolicy by discretizing the covariates in FX before passing them to subpolicy.`