Simulation

ContextualBandits.simulation_stochastic — Function

simulation_stochastic(reps, FX, n, T, policies, outcome_model;
    FXtilde=FX, delay=0, post_reps=0, pilot_samples_per_treatment = 0,
    Xinterest=zeros(length(FX),0), rng=Random.GLOBAL_RNG, verbose=false)

Simulate trials and record metrics.

Arguments

reps: Number of replications.

FX: Covariates generator for in-trial covariates (e.g., CovariatesIndependent or CovariatesCopula).

n: Number of treatments.

T: Number of patients in the trial (sample size).

policies: Dictionary of policies to simulate. The keys are the names of the policies and the values are the policies themselves.

outcome_model: the model that generates the outcomes (e.g., OutcomeLinear).

Optional arguments

FXtilde: Covariates generator for post-trial covariates (e.g., CovariatesIndependent or CovariatesCopula).

delay: Delay in observing outcomes.

post_reps: Number of replications for post-trial covariates.

pilot_samples_per_treatment: Number of pilot samples per treatment to build a prior distribution before the start of the trial. Default is 0.

Xinterest: Covariates of interest for which specific . Default is an empty matrix.

rng: Random number generator. Default is Random.GLOBAL_RNG.

verbose: Print progress of the simulation. Default is false.

Returns

A dictionary with the input arguments and the output of the simulation for each policy.

The return value is a dictionary with two keys: "input" and "output". The "input" saves the input parameters to the simulation. The "output" contains one key for each policy. For each policy, we output the following metrics:

"regret_on": online regret

"cumulregret_on": cumulative online regret

"PICS_on": online PICS

"cumulPICS_on": cumulative online PICS

"Wfrac_on": fraction of allocations to each treatment during trial

"regret_off": offline regret

"PICS_off": offline PICS

"Wfrac_off": fraction of allocations to each treatment after implementation

Online metrics are recorded for each time t=1:T. All offline metrics are recorded for each time t=0:T. Offline metrics at timet represent the metric in the scenario where the trial would have concluded at time t. For the eight online and offline metrics, the output also provides those metrics for each of the specific covariate values specified in Xinterest, which are saved under the same names with an X at the beginning, for example, "Xregret_off" measures the offline regret for subjects that have the covariate values specified in Xinterest.

Two additional outputs are provided that are useful for understanding policies that infer the labeling:

"labeling_frac": Which entries of labeling where true when the trial concluded, which after aggregation represents the fraction of replications in which the entries labeling were true.

"sum_labeling": The number of true entries of labeling.

Parallel Computing

Because simulations can be computationally expensive, the simulation_stochastic_parallel is used to distribute the replications evenly among all available workers (see the standard library Distributed Computing page on how to create worker processes).

ContextualBandits.simulation_stochastic_parallel — Function

simulation_stochastic_parallel(reps, FX, n, T, policies, outcome_model;
    FXtilde=FX, delay=0, post_reps=0, pilot_samples_per_treatment = 0,
    Xinterest=zeros(length(FX),0), rng=Random.GLOBAL_RNG, verbose=false)

Parallel version of simulation_stochastic. The simulation is distributed among all available workers.

Example

using Distributed
addprocs(2)
@everywhere using ContextualBandits
# ...
# generate all the input arguments for the function
# ...
results = simulation_stochastic_parallel(FX, n, T, policies, outcome_model)

An effective workflow is to run parallel simulations in a script. At the top of the script, you need to load the package in all workers with @everywhere:

# example_parallel.jl
using Distributed
@everywhere using ContextualBandits
# ...
# generate all the input arguments for the function
# ...
results = simulation_stochastic_parallel(FX, n, T, policies, outcome_model)

Then you run the script with the -p flag indicating the number of workers you want to use or auto to use all available cores (command line interface for Julia):

julia -p auto example_parallel.jl

If you installed ContextualBandits in a local environment, you also need to activate the environment in all workers using Pkg:

# example_parallel.jl
using Distributed
@everywhere using Pkg
@everywhere Pkg.activate(".") # path to the environment
@everywhere using ContextualBandits
# ...
# generate all the input arguments for the function
# ...
results = simulation_stochastic_parallel(FX, n, T, policies, outcome_model)